# OpenAPI/OpenAPI.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Spring boot template",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/healthcheck": {
      "get": {
        "tags": [
          "healthcheck"
        ],
        "description": "Healthcheck endpoint",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/account": {
      "post": {
        "tags": [
          "account"
        ],
        "description": "Creates new account and stores it into database",
        "requestBody": {
          "description": "account to be added",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "contentSchema/requestModel.json"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "responses/201AccountCreated.json"
          },
          "400": {
            "$ref": "responses/400AccountBadRequest.json"
          },
          "409": {
            "$ref": "responses/409AccountConflict.json"
          },
          "default": {
            "$ref": "responses/default.json"
          }
        }
      }
    },
    "/api/v1/account/{account-id}/role": {
      "post": {
        "tags": [
          "role"
        ],
        "description": "Assign or revoke role",
        "parameters": [
          {
            "name": "account-id",
            "in": "path",
            "description": "Id of account to assign/revoke role",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Account role and status",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "contentSchema/accountRoleRequestModel.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "responses/200AccountRoleOK.json"
          },
          "201": {
            "$ref": "responses/201AccountRoleCreated.json"
          },
          "400": {
            "$ref": "responses/400AccountRoleBadRequest.json"
          },
          "404": {
            "$ref": "responses/404AccountRoleNotFound.json"
          },
          "default": {
            "$ref": "responses/default.json"
          }
        }
      }
    },
    "/api/v1/account/{account-id}/role/{role}": {
      "get" : {
        "description": "Returns account role id for given account id and role",
        "tags" : [
          "role"
        ],
        "parameters": [
          {
            "name": "account-id",
            "in": "path",
            "description": "Id of account to get account role id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "role",
            "in": "path",
            "description": "Role of account to get account role id",
            "required": true,
            "schema": {
              "$ref" : "contentSchema/enums/roleEnum.json"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "responses/200AccountRoleIdOK.json"
          },
          "404": {
            "$ref": "responses/404AccountRoleNotFound.json"
          }
        }
      }
    }
  }
}